feat: Move node exec command to nodegizmo's root
