+++ pkg/cmd/nodes/nodes.go
@@ -44,7 +44,6 @@ func NewCmdNodeInfo() *cobra.Command {
 
 	// additional sub-commands
 	cmd.AddCommand(NewCmdNodeCapacityInfo())
-	cmd.AddCommand(NewCmdNodeExec())
 
 	return cmd
 }
